Extract functions to module
